requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
curl_cffi>=0.7.0
playwright>=1.40.0
google-genai>=1.0.0
